Multiple Characters and fixed Event Naming
Testing for multiple characters.
So we kill one and a time and make sure that we quit the game once
everyone is dead.
This will help when going into production of the real thing because we
can have x amount of characters and the game waits till they are all
dead.